+++ package.json
@@ -22,7 +22,12 @@
   "homepage": "https://github.com/mattcarpowich1/inventory-scrape#readme",
   "dependencies": {
     "@types/puppeteer": "^5.4.4",
+    "apollo-server": "^3.6.3",
+    "apollo-server-express": "^3.6.3",
+    "dotenv": "^16.0.0",
     "express": "^4.17.2",
+    "graphql": "^16.3.0",
+    "graphql-request": "^4.0.0",
     "puppeteer": "^13.3.1"
   },
   "devDependencies": {
